add a force flag to silence the dialogue before clobbering

+++ was.go
@@ -9,28 +9,30 @@ import (
 )
 
 var verbose bool
+var force bool
 
 var ext string = ".was"
 
 func init() {
+	flag.BoolVar(&force, "f", false, "clobber any conflicting files")
 	flag.BoolVar(&verbose, "v", false, "verbose output")
 }
 
 //swiped this from a gist:
 //https://gist.github.com/albrow/5882501
 func askForConfirmation() bool {
-        consolereader := bufio.NewReader(os.Stdin)
+	consolereader := bufio.NewReader(os.Stdin)
 
-        response, err := consolereader.ReadString('\n')
+	response, err := consolereader.ReadString('\n')
 	if err != nil {
 		fmt.Printf("%v\n", err)
 		os.Exit(2)
 	}
 	okayResponses := []string{"y", "Y", "yes", "Yes", "YES"}
 	nokayResponses := []string{"n", "N", "no", "No", "NO"}
-	if containsString(okayResponses, response[:len(response) - 1]) {
+	if containsString(okayResponses, response[:len(response)-1]) {
 		return true
-	} else if containsString(nokayResponses, response[: len(response) - 1]) {
+	} else if containsString(nokayResponses, response[:len(response)-1]) {
 		return false
 	} else {
 		fmt.Println("Please type yes or no and then press enter:")
@@ -48,7 +50,7 @@ func posString(slice []string, element string) int {
 	}
 	return -1
 }
- 
+
 // containsString returns true iff slice contains element
 func containsString(slice []string, element string) bool {
 	return !(posString(slice, element) == -1)
@@ -84,6 +86,7 @@ Read file list from STDIN
 `)
 
 		flag.PrintDefaults()
+		fmt.Fprintf(os.Stderr, "\n")
 	}
 
 	flag.Parse()
@@ -141,17 +144,22 @@ Read file list from STDIN
 				fmt.Fprintf(os.Stderr, "target is blocked:%s\n", targetFile)
 			}
 
-		        fmt.Printf("There's a file in the way:%s:\n", targetFile)
-		        fmt.Printf("Delete %s? Please type yes or no and then press enter:\n", targetFile)
-                        if askForConfirmation() {
-			  if err := os.RemoveAll(targetFile); err != nil {
-				fmt.Fprintf(os.Stderr, "could not clear the way for new was file:skipping:%v\n", err)
-				continue
-			  }
-                        } else {
-				fmt.Fprintf(os.Stderr, "user chose to not delete target:skipping:%s\n", targetFile)
-				continue
-                        }
+			if !force {
+
+				fmt.Printf("There's a file in the way:%s:\n", targetFile)
+				fmt.Printf("Delete %s? Please type yes or no and then press enter:\n", targetFile)
+				if askForConfirmation() {
+					if err := os.RemoveAll(targetFile); err != nil {
+						fmt.Fprintf(os.Stderr, "could not clear the way for new was file:skipping:%v\n", err)
+						continue
+					}
+				} else {
+					fmt.Fprintf(os.Stderr, "user chose to not delete target:skipping:%s\n", targetFile)
+					continue
+				}
+
+			}
+
 		}
 
 		if verbose {
